Replace user/session migrations with players and OTP
