--- server/migration/src/m20240101_000006_create_sessions.rs
@@ -1,83 +1,0 @@
-use sea_orm_migration::prelude::*;
-
-#[derive(DeriveMigrationName)]
-pub struct Migration;
-
-#[async_trait::async_trait]
-impl MigrationTrait for Migration {
-    async fn up(&self, manager: &SchemaManager) -> Result<(), DbErr> {
-        manager
-            .create_table(
-                Table::create()
-                    .table(Sessions::Table)
-                    .if_not_exists()
-                    .col(
-                        ColumnDef::new(Sessions::Id)
-                            .big_integer()
-                            .not_null()
-                            .auto_increment()
-                            .primary_key(),
-                    )
-                    .col(ColumnDef::new(Sessions::UserId).big_integer().not_null())
-                    .col(
-                        ColumnDef::new(Sessions::SessionToken)
-                            .text()
-                            .not_null()
-                            .unique_key(),
-                    )
-                    .col(
-                        ColumnDef::new(Sessions::CreatedAt)
-                            .timestamp_with_time_zone()
-                            .not_null()
-                            .default(Expr::cust("NOW()")),
-                    )
-                    .col(
-                        ColumnDef::new(Sessions::ExpiresAt)
-                            .timestamp_with_time_zone()
-                            .not_null(),
-                    )
-                    .foreign_key(
-                        ForeignKey::create()
-                            .name("fk_sessions_user_id")
-                            .from(Sessions::Table, Sessions::UserId)
-                            .to(Users::Table, Users::Id)
-                            .on_delete(ForeignKeyAction::Cascade),
-                    )
-                    .to_owned(),
-            )
-            .await?;
-        manager
-            .create_index(
-                Index::create()
-                    .name("idx_sessions_user_id")
-                    .table(Sessions::Table)
-                    .col(Sessions::UserId)
-                    .to_owned(),
-            )
-            .await?;
-        Ok(())
-    }
-
-    async fn down(&self, manager: &SchemaManager) -> Result<(), DbErr> {
-        manager
-            .drop_table(Table::drop().table(Sessions::Table).to_owned())
-            .await?;
-        Ok(())
-    }
-}
-
-#[derive(Iden)]
-enum Sessions {
-    Table,
-    Id,
-    UserId,
-    SessionToken,
-    CreatedAt,
-    ExpiresAt,
-}
-
-#[derive(Iden)]
-enum Users {
-    Table,
-    Id,
-}
